third_party/blink/web_tests/pending_beacon/pending_beacon-deactivate.html

<!DOCTYPE html>
<head>
<meta charset="utf-8">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
</head>
<body>
<script>
  test(() => {
    const beacon = new PendingBeacon('https://www.google.com',
                        {method: 'GET', pageHideTimeout: 200});
    assert_true(beacon.isPending);
    beacon.deactivate();
    // TODO(https://github.com/WICG/unload-beacon/issues/11): Update once the issue is resolved.
    assert_true(beacon.isPending);
  }, "Tests deactivating a pending beacon");
</script>
</body>
</html>
